Week 14 commits
https://youtu.be/h_VuNKqWA4U

This was a rough one. I changed the parent version number per instructions in Slack and it deleted all my interfaces and enums along with several packages. It was a real deep dive to get back what I could and remake what I couldn't find. But once I got to the end of Dr. Rob's last Week 14 video Deterministic Testing with an In memory Database (for the 2nd time), everything fell together and I've never felt a hit of dopamine like it my 35 years on Earth. Go me!